requirements.txt
ultralytics
torch
torchvision
opencv-python
onnxruntime-gpu
onnx
numpy
wandb
matplotlib